[Upload] This -> Version
This is the final try. If this works i will go to eat. lol, if my server turned 24/7 i will not try to publish about it. Just me and my friend will only play in that server. If others join, i wont kick them. But if they are rude and bad then i will ban them. Guess nothing bad happends on the jurney of muth. He will be there in that server in the cloud for make us happy. To keep the server alive. Hope this works. Server started now its time for try. Bye - Robin